Use standard normal distributions to classify
calculate Z scores  for an object in each standard distribution for each object. Classified as the lower the type with the lowest score. This is the most general method, and works super well (and only gets better with more data), but is a bit computationally inefficient since it uses floats off the bat. Could be combined with testing for inclusion in the range as a first filter, but that does impose a computational penalty if it wasn't in the range.